Shuffle agents to send request
For each polling interval, Ceilometer compute agent will trigger
nova client directly to acquire the list of VMs.

However, there's a potential risk that the workload for nova client
will increases proportional to the number of Ceilometer compute
agents. In other word, there's a possibility that thousands of
ceilometer compute agents will call nova client in a very short time
period which may lead large number of requests to nova client in
short time. As a result, the nova client can not handle so many
requests in short time, and ceilometer compute agent can't fetch
metrics.

So we here shuffle agents to start polling task in order to fluff up
the time of sending request to nova or other components to avoid
large number of requests in short time. This would be more
performance effective and reliable.

DocImpact

Change-Id: I12e3f104fc92fe15adc05e2b981627f31ee5bfaa
Closes-bug: #1412613