Use stevedore to load alarm rules api
Alarm rule evaluators are loadable with stevedore, but not the API side,
since wsme 0.6 it's now possible to do it.

This change adds that to allow to create a gnocchi alarm rule api and
evaluator in the gnocchi code tree.

This also move out the V2Controller API from __init__ to avoid
recursive import with stevedore.

Change-Id: I468f9f82b41eb7f4896dd93a8c62ce63b49535bb