Initial gabbi testing for alarms
Initial testing of the alarms-related resources in the API. Note that
these tests were initially started by processing the WADL files at
https://github.com/openstack/api-site/tree/master/api-ref/src/wadls
to generate a stubbed out YAML file which was then filled in and
reordered as required. It doesn't save a great deal of time but does
provide a base to start from, which is useful scaffolding to get
started.

Thus, the endpoints in the YAML file are only those endpoints
mentioned in the WADL. If the documentation is not up to date, then
neither are these tests.

The course of these tests is as follows:

* list the alarms confirm there are none
* create an alarm
* list the alarms confirm there are one
* update an alarm
* check an alarm's history
* change state on alarm
* check the state
* delete the alarm
* list the alarms confirm there are none

This is not sufficient for coverage but does provide a good
representation of what's going on in the alarms API.

Tests with 'xfail: true' will require changes either in frameworks or
controllers, pending discussion.

Change-Id: I814b65abb753c4882b0aed48f0b20d6994fb1f30