Enable collector to requeue samples when enabled
If an exception is reported by the storage layer,
the dispatcher logs the error but does not
propagate it. Stevedore plugin framework also does
not propagate the exception back to the caller.

The sample is not requeued since collector does
not get the exception and returns a Handled
NotificationResult. This fix raises the exception
in the database dispatcher and enables Stevedore
to propagate exceptions to the collector.

Change-Id: I341ad92df050bf0f474b07c3f27c68cdb4f4e59b
Closes-Bug: 1433881