Further mock adjustments to deal with intermittent failure
When the py-mysql tox target is run after the py27 tox target it can
sometimes fail because a MagicMock was not properly initialized.
This change more completely fleshes out the mock so that a TypeError
does not occur when a debugging log message is caused. The
specifics of why the logging only happens sometimes are not
understood.

Change-Id: I99f80d794b757a0ac05f7a9b53974ba9913e6e68
Closes-Bug: #1438300