TraitText value restricted to max length 255
Trait_text has a sqlalchemy model definition with
value column as Text type, but the migrate scripts
defined the column type as String(255) leading to
Sqlalchemy exception when indexing trait with large
text.

This fixes the TraitText model value as String(255)
to match the migrate script definition and crops
the text input to 255 characters.

Change-Id: I55eee4e9fcdee8d05f3598a9bc0368bc92a9b238
Closes-Bug: 1447647