Improve ceilometer-api install documentation
This updates the mod_wsgi configuration and installation information
to be more in line with current usage. It also makes the
recommendation to use mod_wsgi more strongly than has been made in
the past. In general we don't want people to be running
ceilometer-api at all, but it is good for them to have the option.

Also an effort has been made to clarify the difference between rpm
and deb installations of Apache.

DocImpact

Change-Id: I11f92d91f4bdea884c266b91b52f9d012abd33b9
Partially-Implements: blueprint remove-web-eventlet