VMware: verify vCenter server certificate
Two configuration properties are being added:

'ca_file': Specify a CA bundle file to use in verifying the vCenter
server certificate

'insecure': If true, the vCenter server certificate is not verified.
If false, then the default CA truststore is used for verification.
This option is ignored if 'ca_file' is set.

Closes-Bug: #1276207

DocImpact

Change-Id: I8f408308cddbb40b19e8dc9fce6ff02745d963b8