[elasticsearch] default trait type to string
when trait mapping is not found, get_events for es backend will
assume the last known type that was mapped. this patch changes that
so it will default to text type if a mapping is not found.

Change-Id: I5c5b263f0160726034dfca59d70d2a6a32585cb5
Closes-Bug: #1453919
(cherry picked from commit 906c72fc216a108cd949faf94fc2a2adda014a9b)