Handle database failures on api startup
Storage layer connections are initialized
for metering, alarm and event in Pecan hook
middleware.
If any of the backends are not responding,
the connections are retried till maximum
retries and the exception propagated to the
Ceilometer-api which fails to startup.

This fix logs and handles the exception in
the Pecan hook and allows all connection
types to be initialized.

Closes-Bug: 1456944

Change-Id: Id70693dafc823927b3d72e0371e3e55d4e5588ce
(cherry picked from commit 1f30839494c5b5002173a1f81eb960d735f39643)