Fix usage of dictionary methods on Python 3
Convert the result of keys(), values() and items() dict method to list
when a list is expected. On Python 3, these methods now return an
iterator, not a list.

Instead of getting the key with .keys()[0] and then the value with
.values()[0], get both at the same time using .items()[0]. It ensures
that the key is consistent with the value, but it is also more
efficient.

Enable ceilometer.tests.storage.test_impl_sqlalchemy on Python 3.4.

Change-Id: I4bcd3dfbca3e26238cdcdf19801244497dac2729