Fix testing of agent manager with tooz
The test test_load_plugins_pollster_list_forbidden is currently failing
because the agent tries to load the tooz driver before doing the sanity
check. That makes tooz trying to load an http driver that does not
exist, and the test fails.

Let's do the check before anything else to fix that.

Closes-Bug: #1463636
Change-Id: I646ac0ed0fb9db04fa81f10cd8da0ac7bff5c138