Restricts pipeline to have unique source names
Sample and event pipelines are validated for unique
sink names and unique pipeline names(combined source
and sink names), but not unique source names.

This fix validates source names to be unique in
pipeline.

Also refactors the logic to validate sink and pipe
names for duplicates using set instead of list
iteration.

Change-Id: I349b48fc5e50f1bb16e453a0052021b855230ccc
Closes-Bug: 1466136