Make ceilometer work correctly when hosted with a SCRIPT_NAME
In the past ceilometer was generaly hosted at / and URL generation
was relatively straightforward. It should also be possible to host
at /foobar or /telemetry and for URL generation to still work. This
changeset adds tests to confirm prefixed requests can work and fixes
the small number of issues that were found.

Note that because the ceilometer api has the temerity to not set
location headers, we can't test that those are working, because
there aren't any. In general that's the place where SCRIPT_NAME
mounting would show bugs.

But we can confirm that content-location headers are okay.

Closes-Bug: #1468796
Change-Id: I6e5cf440bc3bd8ffb4cc5c86aa19225b3c5f87ed