Grenade plugin using devstack plugin for ceilometer
This is barebones for now, simply doing the upgrade and checking
services are still running. There is no checking of resources.
Being able to easily add those resources checks is part of the
reason for making this change.

The other reason is to explore and resolve conflicts when extracting
an existing integrated project from both devstack and grenade.

Note that grenade's ensure_services_started and _stopped struggles to
work with processes that contain spaces so distinguishing amongst
ceilometer-polling processes with different namespaces is hard, so
we'll just use the older agent names.

Change-Id: If7bb809743ab0c475a1844c79e552d8b6892a18f