Drop out-of-time-sequence rate of change samples
Now that samples always go over the messaging bus before being
potentially supplied to the transformers, there is a possibility
that samples can arrive out of time order. When this happens a
negative time delta could happen, causing a negative number where
one makes no sense.

This change adjusts handling such that if a sample if older than the
sample in the cache it is dropped and the cache is reset to have the
most recent sample. A dropped sample is logged with a warning as we
assume that violations of the order of time are something we might
want to be aware of.

If the samples are the exact same age, we carry on.

Change-Id: I8c8fe0c0ab3c01ef5b2913ab18dfd5959aefaed3