Remove the compute inspector choice restriction
A recent change ( https://review.openstack.org/#/c/203319/ ) was pushed
into the compute inspector branch to identify the choices for the
varying Ceilometer inspectors.

The side effect of this change is that it limited the operator to only
be allowed to use those options.  Any other option would throw an error.

Given the approach to use plugins under the ceilometer.compute.virt
namespace, the inspectors can now reside out of tree.

It was discussed in the July 30th Ceilometer IRC meeting, and it was
decided instead that the current known options should instead be listed
in the help, rather than limiting via the choices attribute.  This
allows for plugins (ex. the PowerVM compute inspector) to be used.

Change-Id: Ie8f881b7b04372cab06d8c8b61693c5489b2b302