Provide base method for inspect_memory_resident
The inspect_memory_resident was introduced a few months back.  When a
new inspector element is added, it should also be added to the base
inspector.  The base inspector should raise a NotImplementedError.

Due to this missing defintion, the pollster for the resident memory was
throwing an error for every instance that was sampled - if that
inspector did not have this method defined.

This change set adds a base definition, which removes the excessive
logging per instance when using a non-libvirt inspector.

Change-Id: Ibac8a18a3e84fb90763c48c892583d8d3785c91c
Closes-Bug: 1480309