Fixed event requeuing/ack on publisher failure
Event should be acked or requeued when either event
conversion or event publishing fails by configuring
ack_on_event_error in configuration.

When publishing fails, the exception is not raised
in the pipeline and event does not get acked or
requeued, with a Handled message sent to transport
leading to event loss.

This fix raises the exception in pipeline on publish
failure and improves unit tests to test event handling
with a real pipeline.

Change-Id: Ia00a9ce30c5b03269091caa7f5edca9fcf113aff
Closes-Bug: #1482797