Events RBAC needs scoped token
The role based access control of events api needs project-
scoped tokens to process the request. The project information
has to be available in the request header to be able to filter
events based on the project.In addition to the project, the user
information is also mandated as part of this change. The request
fails with a 403 if this information is not available in the
request.

Change-Id: I55e3d8eaf278024e89e43e36a9dbb92b1c432646
blueprint: events-rbac