Correct the host field of instance metadata
Fow now, the 'host' field of instance resource medata is instance.hostId,
which is a hash string based on project id and hypervisor name, it is
hard to indicate which host the instance running on. This change add a new
instance metadata field "instance_host" to indicate the hostname of the
host the instance running on.

Change-Id: Iaa6fa9ef099f617da30cc7af74aa74a845064489
Closes-Bug: #1506610