avoid generate temporary table when query samples
Currently, when we call query-samples API, for SQL backend, we do
an inner join to get a temporary table, then apply filters, orderby
and limit options on it. Such implementation will drain disk space
or cause timeout exception when data set is growing, even with specified
limit and filters.

This patch applies filters and limit on the inner join itself, and
uses a fake mapper object instead of temporary table to suit current
query transformer.

Change-Id: I261a2dd362ed51c16a6fa191dadcce1b45fce2e4
Closes-Bug: #1506738