Adjust log levels for InstanceShutOffException
Logged messages for compute pollsters are brought in line with the
guidance in openstack-specs as noted in the bug report.

NoDataExceptions still produce a warning as that may be due to cloud
configuration and may be actionable for operators.

Unit tests for compute memory pollsters are adjusted to accommodate
when warnings are not logged.

Change-Id: I84824e409f0643f50f196ba89187f864781a3154
Closes-Bug: #1554318