'ceilometer-polling' should fail with no valid pollsters
The command 'ceilometer-polling --pollster-list' should fail with
invalid pollsters.

When executing the command 'ceilometer-polling --pollster-list' with
a wrong pollster like aaa, it still executes and no error will
be shown. It occupies the system resources and shows no errors.

Remove the test case for
"ceilometer-polling --polling-namespaces ipmi" in the upgrade.sh.
In the test environment, the ipmi is not ready. It should fail.

Change-Id: If1f9f9fdbba1832e66d89d6b61f749790df5c2b6
Closes-Bug: #1553926