Remove gabbi tests that check content-location
The content-location header is an artifact of using httplib2
and not something that ceilometer itself returns. Gabbi is going
to stop using httplib2 (and so is much of OpenStack) so to prepare
for that these tests need to stop testing something that doesn't
really mean anything.

Change-Id: If5aacaf108402146d66fee474793dde1c0a59a7c
(cherry picked from commit 639b9476bd8554bee93512a81ae24605c766f080)