Refactor floatingip pollster to use discovery
Note: The bug isn't present in master because it is fixed as a side
effect of the patch that is cherry-picked here. Still it seems better to
use this code than to come up with a dedicated bug-fix patch.

Closes-Bug: 1604328

Change-Id: I47cd14fb6c06ea2c83adab987a0442ef9bd3f404
(cherry picked from commit c262389819cc31c9b87ca5207a809fb2aa1ac882)