[Trivial] Refactor libvirt inspector connection & uri
libvirt inpsector has an attribute called `uri`, which only used
when connecting to libvirtd. There is no such need to do so, a
local variable can do it as well.

Meanwhile, we initialize connection by calling a method
`_get_connection`, and there is an attribute ``connection`` holds
the returned value. This method can be replaced with a property
`connection`, to shorten typing and avoid function calling style.

Change-Id: I8604d00c9298f56009c579fd23adc2c93ad44f8c