start partition coordinator only when there is a group to join
Currently we always start a partition coordinator for polling agent,
but sometimes, we don't need it. For example, the compute polling
agent actually is naturally workload partitioned, hence no need
to do a workload partition again for most cases. So for such
scenario, we should not start a partition coordinator for it.

After we split discover into different namespaces, this becomes
a natural decision.

By doing so, we can:

* reduce a long connection to coordinator backend, such as zookeeper,
  this gains much more when cloud scale grows
* start polling immediately after we finish setup works, no delay is
  required if you don't enable shuffle time for polling, this gains
  more if pollsters' interval is a bit higher, for example, 30 minutes
* reduce a heartbeat thread

Change-Id: I6b836041b34e2c99229d71c8fabad33de127847d