collector: use an intermediate proxy class for event dispatcher
This makes sure that the base class EventDispatcherBase does not owns
any code itself. This makes external plugin easier to implement as they
don't _have_ to inherit from EventDispatcherBase, they just need to
implement a `record_events(events)' method and that's it.

This will allow Panko to provide the database dispatcher without having
a hard dependency on Ceilometer.

Change-Id: I094aed1e7e081fe1de130d987035a6c9d977d25e