Fix the py34 jenkins job
The oslo.messaging<5.2.0(the v5.2.0 update change is in review[1])
depends on amqp>=1.4.0, in amqp-2.0 and amqp2.0.1, there is an issue[2]
that lead the amqp cannot be correctly installed in py34
environment(the depent package vine-1.1.0 of amqp won't bee installed),
that has borken the ceilometer py34 job[3]. we can wait the change[1]
merged and then bump the oslo.message version to avoid this issue. Because
Ceilometer don't participate the global-requirements, so we bump the
oslo.messaging to 5.2.0.

[1] https://review.openstack.org/#/c/324023/
[2] https://github.com/celery/py-amqp/issues/91
[3] http://logs.openstack.org/41/323641/4/check/gate-ceilometer-python34/cf3f043/console.html

Change-Id: Icb34a7f0e03662868aa0ce6e060c5e951711b79e