Fix wrong parameter reference in periodic decorator
When setting up polling tasks with different interval, peroidic
decorator refers one outside variable, polling_task. However, this
happends in For loop, so that polling_task is constant across all
iterations and all intervals are related to last polling_task. This
patch adds extra parameter for the decorator function to fix it.

Change-Id: Id3be9078a9b9c8e0598cd7ddc88ccd02d6665d16
Closes-Bug: #1606852