Remove store_events options
The current option is misnamed, as it does not enable any storing but
just the processing of events from the notification agent to the
collector.

This means that even if you set event_dispatchers=panko and forget to
set store_events=true, nothing will happen.

This patch enable the event processing as soon as something is
configured in the pipeline.

Change-Id: I5a906684f6371b0548ac08cacc13aa238f780f78