Add http_proxy_to_wsgi to api-paste
This sets up the HTTPProxyToWSGI middleware in front of Ceilometer. The
purpose of thise middleware is to set up the request URL correctly in
case there is a proxy (For instance, a loadbalancer such as HAProxy)
in front of Ceilometer.

So, for instance, when TLS connections are being terminated in the
proxy, and one tries to get the versions from the / resource of
Ceilometer, one will notice that the protocol is incorrect; It will show
'http' instead of 'https'. So this middleware handles such cases.
Thus helping Keystone discovery work correctly.

The HTTPProxyToWSGI is off by default and needs to be enabled via a
configuration value.

Change-Id: I24f16dda49bd9e7930ca9f0d32bf0793463aff03
Closes-Bug: #1590608
(cherry picked from commit 30eb0f0ae16c7ffd0f90c3874d3cbc462d5863a0)