Revert "Add hypervisor inspector sanity check"
This change have introduced an issue, when the libvirt unix socket
is created after Ceilometer start. The pollster is disabled when
it should wait that libvirt come up.

The intent of the original bug was only to have a cosmetic logs
when the configuration file is wrong.

Also I don't see a safe way to disable the pollster when we can't
connect to libvirt, the reason is always unknown from Ceilometer PoV.

This reverts commit d17e75d2675c1acf09adc27249b1e968450c5229.

Change-Id: I63479d19fcf4a99fb198859c2e6ae370ff0e24e1