pipeline: remove tests helper from runtime code
cfg_info must not be a dict, so remove the code than handle the
that. This also refactor a bit to remove variable shallowing from
load_config().

The tests now mock the file.

test_unit_identified_source_unit_conversion is now skipped since it
have encoding issue when pipeline is read from a file.

Change-Id: I276607b9836a358d5b183d7097b27cb4cefbd684