support gnocchi timeout
set gnocchi timeout to 6.05 because requests recommends multiple
of 3[1]. also, it seems safe to assume gnocchi is not functioning well
and will not scale if we can't post in 6s. in my benchmark, it took
under 0.25s to post 5000 points. the only thing that took >3s was
getting >80K datapoints and this workflow is only pushing, and only
getting small chunks occasionally. making it 6s because it seems
keystone adds decent overhead.

do not set timeout for upgrade because it might take longer than normal
workflow

[1] http://docs.python-requests.org/en/master/user/advanced/#timeouts

Change-Id: I41b5e4f0c0f54b7f4b627d9116e63096ca2decf9
Closes-Bug: #1666072