Don't run ceilometer-upgrade on unconfigured db
When the database backend is unconfigured, we currently
try to connection to nothing, and due to the retry logic around
the database connection, the user experience is "ceilometer-upgrade is
stuck".

This change keeps the database upgrade is the database is not
configured. And don't use the connection retry logic for upgrade.

Change-Id: I3fa2a332ecac6cb52c145f437f79b2aba43a4757