Remove pipeline periodic refresh feature
The pipeline dynamic refresh code has been removed. Ceilometer relies on the
cotyledon library for a few releases which provides reload functionality by
sending SIGHUP to the signal. This achieves the same feature while making sure
the reload is explicit once the file is correctly and entirely written to the
disk, avoiding the failing load of half-written files.

Change-Id: I1aca9cf4c3634e7c70eea651e1dd5c2f3ebfecc6