tests: simplify broken test
The current reset listener on refresh test does not test anything. What it does
is to check that they are 20 oslo.messaging calls, and then if you call the
refresh function, they are called again. Which can be assumed by simply looking
at the code. The test is currently broken because due to workload partitioning
being enabled, it can be different than 20.

This changes it back to its original behaviour, which is checking the listener
is refreshed.

Change-Id: I2354e3e08000077ce867daff3480b57469e15985