tempest: remove confusing stack check step
We look at the heat event to see if the stack is created in less than
600 seconds. And then check the state is STACK_CREATED. But if the setup
is slow (like in rdo/puppet jobs) the state after the event occurs is
STACK_CREATE_IN_PROGRESS and 5 seconds can be not enough to have the
stack fully created.

The save occurs for the update.

This change removes the event checking and just wait 600 for
STACK_CREATED/STACK_UPDATED.

Change-Id: I3ef2898ce61227c62eb02fedeec9ea429d35aa89