stop converting to old message format
process notifications as is. this should not impact upgrades as
this does not change what is outputed by notification agent in
any way, just how it does initial processing.

note, this *will* change what is stored in Events if raw field is
enabled. that said, the raw field is unqueryable from Panko pov and
it actually is storing the raw message now.

Change-Id: If79f91687c6c1bc27d906045573674b002534ab1