Revert "deprecate archive policy of Gnocchi dispatcher"
This reverts commit 5016d7e26ae03d3e427b73c01f93c514612e61f0.

I don't think it's a good idea to deprecate that settings. They are some weird
use cases that are not covered by archive policy rules, such as posting from 2
Ceilometer with different archive policies to a single Gnocchi instance.

There's also the possibility that the Gnocchi instance is used for some other
means and that the rule cannot correctly be set.

Change-Id: I4ee646088f7bc370ed7ffd1eb3b98e66bae6782e