Deprecate kafka publisher
Original kafka publisher has some restrictions on demand. oslo.messaging
has introduced kafka driver and make it scalable, so it's recommended to
integrate kafka driver with NotifierPublisher and push events to broker.
And this patch also implements a generic NotifierPublisher to push data
to external server with customized transport driver.

Closes-Bug: #1649947
Change-Id: I3eecd74b61a42ba1c760b53032ac6e80e85d1dda