Use tempest.clients.Manager as base clients class
The current client Manager uses tempest.manager.Manager, which is
fine for all clients except object storage, since the object storage
client is not yet a stable interface, and it's being moved from
tempest namespace to tempest.lib namespace.

tempest.clients.Manager is a stable interface (not officially yet
but it will be declared so soon, and there are no changes pending).
This class has the advantage of masking the move of the object
storage client providing a smoother experience for ceilometer
through this transition.

The client aliases are consistent with those used by ceilometer,
so the change is very small and it simplifies the code.

Once the object storage client is fully migrated, it would be nice
to expose the Ceilometer service client via the plugin interface,
so it can be loaded automatically and it can be used by other
projects as well.

Change-Id: If71c7bb193fe562ea482e1031c729c789c761bd8
(cherry picked from commit bca9d45ea439220305539f85a653cb31e501be25)