remove sample sorting
this does nothing. the listener currently consumes events across
all queues, drops them into an 'incoming' queue. from here, a thread
batches them based on batch options. the batched messages themselves
are still in order and therefore sorting it does nothing.

the only reason it'd be out of order is if a message was requeue'd.
in this case, sorting will probably do nothing since the requeue'd location
is probably not remotely close to original ordering and will probably
never be batched with same messages to sort correctly.

Change-Id: I6faa97bb90bee00ec94f4faf622822ea93655efb