Remove shuffle_time_before_polling_task option
The problem that shuffle_time_before_polling_task tries to solve is the startup
of a horde of Ceilometer instances that would start polling the same thing at
the same time.

It's actually unlikely they would all start at the same right second, and the
correct fix would be to do that each time.

Related-Bug: #1734898

Change-Id: If8141f6b48657c06e8e782eeef9b209dabb2097c