ignore compute.instance.create.start for metrics
do not build metrics off compute.instance.create.start as it
doesn't have enough information in it to push to Gnocchi. there is
an .end event that happens seconds later so .start event has little
value outside context of event.

we change fnmatch to regex for more useful regex comparison and it's
faster

not a complete fix as compute.instance.update also is missing host
value occasionally

Partial-Bug: #1718290
Change-Id: I2aa7045873fd740255a3be5f64b9e9f5fd94e2d4