ignore compute.instance.update for metrics
nova sends *.update events in between create.start and create.end
events and it might not have a host name (which we need). in reality
to any changes to vcpus/memory/root_gb/ephemeral_gb metrics will
either send a *.resize notification and/or require a reboot which
sends it's own notification. because of that, just ignore
*.update notifications

Change-Id: I1f975c2d73adfc19126d70bdb1ab452e37b278fd
closes-bug: #1718290