Gracefully handle missing metadata in libvirt xml
Missing metadata in the libvirt domain xml for a nova instance would
cause ceilometer-compute to abort, leading to missing metrics for the
current and any subequent libvirt guests.

This commit puts a try/except AttributeError block around all the code
that is fetching attributes on the result of metadata_xml.find(...).

Change-Id: I8adc609cc21c86de2daba326d24b73a80d6eb61f
Closes-Bug: #1749960