Don't use gnocchiclient during publisher init
We can't create archive policies in publisher init because if it fail
because Gnocchi is not yet ready the publisher fail to load. Ceilometer
will work, but samples will go to nowhere.

This change creates/checks archive policies when we publish a sample for
the very first time. If that fail because Gnocchi is not ready, this
will just retry next time Ceilometer will publish samples.

Closes-Bug: #1752420

Change-Id: Ib6b4da54592ad99a4e6561a73473b6c7ec73a21f