objectstore/rgw: Add config option to support RGW implicit tenants
If "rgw keystone implicit tenants" is enabled for Rados Gateway to
enable Multi-Tenacy [1], the created RGW user IDs are in the format
"keystone_project_id$keystone_project_id", i.e. the RGW tenant ID is
added to the user ID. This breaks metering via Ceilometer as the usage
is always queried for uid=keystone_project_id. If implicit tenants are
enabled in RGW, Ceilometer has to query based on the adapted used IDs.
This commit introduces support for querying the correct user accounts.

[1] http://docs.ceph.com/docs/mimic/radosgw/multitenancy/

Change-Id: I6de4c5ce0e4f59c1d952f8fd39de64937e781280
Signed-off-by: Felix Walter <felix.walter@cloudandheat.com>