pipeline: remove transformer support
Transformers cannot work correctly on multiple nodes without workload
partitioning, which has been removed.
The transformation of data is no more the responsability of Ceilometer. The
data storage used by default (Gnocchi), is able to handle the scenario that
Ceilometer covered by default such as computing the rate of a metric.

Change-Id: If3683318b998a37c40bc09314dd8ab3eef326ba7
Depends-On: Ifd1d04ce813028f115c19bc983e7dd1e63c6f8a5
Depends-On: I0330c09d72c20d63d08770b52d3071512a418260