Only print polling.yaml file contents as DEBUG
Since this file can contain sensitive information such
as admin credentials of target resources (e.g. snmp) it
is not a good idea to print this log as INFO since admins
might want to have INFO level logging but without the risk
of it containing this kind of information.

We might want to consider remmoving this log altogether.

Change-Id: Iff5fcc4b6c30142f2c5c9ad26027dc5a6542a881
Closes-Bug: #1811098