Change sample id from int to BigInt
For long running users of Ceilometer with heavy sample loads,
the index for samples can run out of numbers.  The int type only
allows over 2 billion samples to be loaded.  To allow a user
with database backed sample storage (postgres or mysql) to continue
to operate, we need to up the data type to BigInteger.

Change-Id: I9ed1d69a05efd47368497674ffe68b43c8dc1074