Remove _ceilometer_check_for_storage check
We can check from the history[0], actually it doesn't matter if
gnocchi or panko is started first., so this check is unnecessary.
Also when i want to migration using native zuulv3, the enabled
plugins is in the alphabetical order, so this will caused can not
install devstack correctly.

[0]: https://review.opendev.org/#/c/387620
[1]: http://logs.openstack.org/99/651999/3/check/telemetry-integration-test/9066108/controller/logs/devstacklog.txt.gz#_2019-04-17_07_26_30_988
Change-Id: I5cbd93739f75b8141f8b79ae03bb8b99756b2a38
(cherry picked from commit 27eb3f431e177054789ef47859517d10862bc902)