Dynamic pollsters to handle different response structures
This PR addressed a concern raised by Lingxian Kong on
https://review.opendev.org/#/c/677031. The dynamic pollsters were using
just the first key to the response that is received when we call an
OpenStack API. However, the operator might call an API that returns
directly the list of elements (e.g. AODH alarms). Therefore, we want to
enable operators to handle those APIs as well. Moreover, we are
improving the system and allowing operators to define the key of the
list in the response that he/she wants to handle.

Change-Id: Ie0439f5555c1fbc651632bb84040ca794b203139
Signed-off-by: Rafael Weingärtner <rafael@apache.org>