Add support to linked samples responses
Some APIs split their responses into pages and each response has
a link to the next one.

To allow operators to deal with this kind of APIs, we propose to
extend the Dynamic pollsters to navigate through the API
responses and join all responses into a single one.

To enable it, the operator will need to configure the parameter
`next_sample_url_attribute` with a mapper to the response's next
page attribute.

Change-Id: Ida0a73d2964f192e6c63a6b7e8003ef2b52bd710