Tell reno to ignore the kilo branch
When reno 3.x runs under setuptools, it scans all of the branches it can
find, including any that look like they're closed and have an -eol tag.
The old kilo branch in this repository has a jumbled history that
somehow makes it look like it should include tags that it doesn't. We
know that there are no release notes in that branch, because reno wasn't
adopted while it was open. The releasenotes/source/index.rst links to
separate release notes in the wiki.

This patch tells reno to ignore that branch so that it doesn't throw an
exception when it gets confused about the old tag.

Change-Id: I8047d805b0a3da4051a6d618ed9ac40875b8636a
Co-Authored-By: Doug Hellmann <doug@doughellmann.com>
Story: 2007274
Task: 38749