Dynamic pollster support paging systems that return only a path for the next page
This was found when working with Glance API, where the next page attribute only
contains the context for the next call, and not a full qualified web address.
Moreover, while implementing this one, I noticed the there was no paging for
the non-openstack pollster. Therefore, I added it as well.

Depends-On: https://review.opendev.org/#/c/709807/

Change-Id: I3bc3cd4a7aa58ca1e2a57d383933493f5dc377ca