Only process 'resource_id' if metric is configured for Gnocchi push
Somne errors are being caused by the new neutron implementation of network metering
(https://review.opendev.org/#/c/735605/). The samples there (in the new implementation)
use a different format than the expected by the default "bandwith" and "packages" meter
notification configuration (in ceilometer notification).

When enabling the new Neutron implementation, we need to remove the following meters
(out of the network resource):
- bandwith
- packages

Those metrics are defined by default in Ceilometer-notification. The problem is that those
definitions expect the old format of Neutron metering messages. Therefore, whenever we enable
the granular extension for Neutron metering system, we need to disable them.

However, even if we disable them, we have a problem. For all metrics (even the ignored ones)
Ceilometer-notification was executing the following:

```
resource_id = resource_id.replace('/', '_')

```

This causes an error because the `packages` and `bandwith` metrics generated with the default
ceilometer-notification will not have a resource_id.

Change-Id: I04e098d3e1fa211ac7cff37681389de50f38197a