vmware tests: Support different moref backend representations
We want to switch the SOAP library backing oslo.vmware [1], which
changes how a ManagedObjectReference's attributes look like. To handle
the two representations in the tests - especially for the transition
period - we will set both attributes on fake.ManagedObjectReference.
Therefore, we have to use fake.ManagedObjectReference everywhere instead
of using manual Mock objects lacking one of the attributes.

[1] https://specs.openstack.org/openstack/oslo-specs/specs/victoria/oslo-vmware-soap-library-switch.html

Change-Id: I449e045e948ad543c4d046b6449094c7617ce78f